Add files via upload
Before implementing cloud connectivity, the system was initially developed as an offline solution, where data was processed and displayed locally on the device. During this phase, the Thonny IDE was used to write and execute Python scripts directly on the ESP8266, allowing us to test and debug the sensor readings in a controlled environment.

Additionally, we built two independent circuits, each equipped with a DHT11 sensor, to measure temperature and humidity separately. These circuits operated autonomously, collecting and displaying real-time data for validation and comparison before integrating the system into the IoT platform.