Add files via upload
The ESP8266 microcontroller collects sensor readings and communicates the data to ThingsBoard Demo, an open-source IoT platform. A Python script was used to facilitate data processing, formatting, and transmission. ThingsBoard was utilized to visualize the readings through a web-based dashboard, allowing for real-time monitoring and analysis. This project demonstrates the application of embedded systems and IoT technology for environmental sensing, offering insights into data-driven monitoring solutions.